requirements.txt
# Core Framework
chainlit==2.9.6

# LangChain Core & Integrations
langchain-core==0.3.28
langchain-community==0.3.13
langchain-classic==0.3.9
langchain-chroma==0.2.1
langchain-huggingface==0.1.2
langchain-groq==0.2.1
langchain-text-splitters==0.3.4

# Vector Database
chromadb==0.5.23

# Embeddings & Models
sentence-transformers==3.3.1
transformers==4.47.1

# Reranking
flashrank==0.2.9

# Environment & Utils
python-dotenv==1.0.1

# Additional Dependencies (auto-installed with above, but explicit for clarity)
pydantic==2.10.5
pydantic-settings==2.7.1
